requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings
groq>=0.4.0
tmdbv3api
rich>=13.0.0
typer>=0.9.0
pyautogen>=0.2.0
numpy>=1.24.0
pandas>=2.0.0 